mptcp: sockopt: ip tos: remove win
The window size has recently been tweaked on the kernel side, see commit
378979e94e95 ("tcp: remove 64 KByte limit for initial tp->rcv_wnd value"),
and caused the IP TOS tests to fail because they explicitly expect a
window of a certain size.

These window sizes for the outgoing packets are optional, and in fact
not used anywhere else in MPTCP tests. Verified using:

  $ git grep '+\S\+\s\+> .\+win [1-9]' -- gtests/net/mptcp/

So remove these size to avoid the issue. These tests are not there to
validate the window sizes, but other features.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>