mptcp: global: slow down connections
Many of the recent instabilities have been due to the TCP stack being
too aggressive to retransmit packets. Adding some delays in the injected
packets at the beginning of the connections often help to avoid too
aggressive retransmissions.

In this commit, instead of looking at errors reported by the CI, the
different files have been manually analysed to add some delays were it
was missing.

While at it, unnecessary delays added to the packets generated by the
kernel have been removed: if there is no reason to have a delay, no need
to add one and expect that the tolerance will save us.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>