mptcp: prio: MPJ+SYN+ACK should not have the backup flag
This 'backup' flag should only be set if the peer sending it asked for
it.

Here, the host didn't ask for it -- no endpoint linked to this address
using with the backup flag -- it is simply mirroring the one set by the
other peer in the MPJ + SYN as a consequence of a bug in the kernel.

Linked to the 'mptcp: fix inconsistent backup usage' series, and more
particularly to the 'mptcp: distinguish rcv vs sent backup flag in
requests' patch.

Link: https://lore.kernel.org/r/20240711-mptcp-backup-mpj-v1-0-d45506182a9e@kernel.org
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>