mptcp: close one SF, expect a FIN back
When a connection has multiple subflows, and one of them is being closed
by sending a FIN, the kernel was not replying with a FIN+ACK back.

This bug has not been seen before, probably because the in-kernel
path-manager always sends a RM_ADDR before closing the subflow, forcing
the closure from the other peer.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>