mptcp: check fallback if opt dropped after first data
With the current kernel, the server incorrectly resets a connection if
its first packet has data (< init window) and MPTCP options, but then
the next ones no longer have MPTCP options. [1]

The server should not reset this connection in this case, but do a
fallback. It should reset the connection only if it has received a
DATA_ACK, or data for more than the initial window (implying DATA_ACK
have been received by the other peer) [2].

This new packetdrill test checks the expected behaviour.

Link: https://github.com/multipath-tcp/mptcp_net-next/issues/518 [1]
Link: https://datatracker.ietf.org/doc/html/rfc8684#name-fallback [2]
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>