mptcp: mpc: no blackhole if not 1st TCP accepted
An MPTCP firewall blackhole is supposed to be detected if only the
following SYN retransmission after a fallback to TCP is accepted.

If it is a different one, the blackhole protection feature should not be
on, and that's what this new test is checking.

While at it, it also checks if the new syn_retrans_before_tcp_fallback
sysctl knob is working as expected.

Link: https://lore.kernel.org/20250114-mpc-no-blackhole-v1-0-994bd2a357fb@kernel.org
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>