mptcp: close_before_accept: remove extra endpoint
This test recently failed [1] because the ADD_ADDR got retransmitted:

  # close_before_accept.pkt:23: error handling packet: live packet field ipv6_payload_len: expected: 44 (0x2c) vs actual: 60 (0x3c)
  # script packet:  7.351175 . 1:1(0) ack 3 <nop,nop,TS val 100 ecr 700,dss dack8 216172782113783808 flags: Aa>
  # actual packet:  7.351166 . 1:1(0) ack 3 win 1053 <nop,nop,TS val 616 ecr 700,add_address address_id: 1 ipv6: fd3d:a0b:17d6::2 hmac: 7019336500874436005>

In this test, the server was configured with an extra endpoint, hence a
first ADD_ADDR sent. But no ADD_ADDR echo got injected. Instead of
adding this echo, it seems better not to add the extra endpoint, and not
have to deal with ADD_ADDR's that are not relevant for this particular
test.

Link: https://github.com/multipath-tcp/mptcp_net-next/actions/runs/17641769831
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>