mptcp: mpc: validate blackhole recover path
Code in mptcp_active_enable() has been recently changed in net-next,
with a few fixes, and one more coming.

I noticed we didn't have any tests covering this path: an MPTCP
connection can be established after the blackhole period. This is what
this test is doing, with a blackhole period of 1 second.

Note that this test doesn't expose an issue directly, but it would let
other debug tools finding some, if any.

Link: https://ci-results.mptcp.dev/lcov/export/mptcp/ctrl.c.gcov.html
Link: https://lore.kernel.org/20250916214758.650211-7-kuniyu@google.com
Link: https://lore.kernel.org/20250916214758.650211-8-kuniyu@google.com
Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>