mptcp: dss: retrans data-fin once more
On slow environments, the test can fail because the DATA_FIN ACK can be
injected with a too long delay after the 'ss' command.

To avoid that, an extra retransmit is expected with a greater tolerance.

While at it, also reduce the delay to inject the DATA_FIN ACK in
dss_fin_retrans_last_ack to avoid a retransmission of the DATA_FIN.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>