mptcp: dss: no time expectation after ss cmd
On slow environments, the test can fail because the DATA_FIN ACK can be
injected with a too long delay after the 'ss' command.

To avoid that, an extra retransmit has been expected with a greater
tolerance. Still, the test has been seen failing with:

  dss_fin_rcv_retrans_fin.pkt:32: error handling packet: timing error: expected outbound packet in relative time range +0.000000~+1.600000 sec but happened at -2.492486 sec

Remove the time expectation for this retransmission.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>